Extract exception tree webview to separate module
